+++ mitmproxy/net/tls.py
@@ -146,7 +146,6 @@ def create_proxy_server_context(
     verify: Verify,
     ca_path: str | None,
     ca_pemfile: str | None,
-    client_cert: str | None,
     legacy_server_connect: bool,
 ) -> SSL.Context:
     context: SSL.Context = _create_ssl_context(
@@ -166,14 +165,6 @@ def create_proxy_server_context(
         raise RuntimeError(
             f"Cannot load trusted certificates ({ca_pemfile=}, {ca_path=})."
         ) from e
-
-    # Client Certs
-    if client_cert:
-        try:
-            context.use_privatekey_file(client_cert)
-            context.use_certificate_chain_file(client_cert)
-        except SSL.Error as e:
-            raise RuntimeError(f"Cannot load TLS client certificate: {e}") from e
 
     if legacy_server_connect:
         context.set_options(OP_LEGACY_SERVER_CONNECT)